Added NotificationService as part of Prism.Avalonia.Toolkit (:
