+++ Prism.Avalonia.sln
@@ -7,6 +7,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Prism.Avalonia", "src\Prism
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Prism.DryIoc.Avalonia", "src\Prism.DryIoc.Avalonia\Prism.DryIoc.Avalonia.csproj", "{A379E2C8-BD9A-42A1-8218-CED3C0D75E64}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Prism.Avalonia.Toolkit", "src\Prism.Avalonia.Toolkit\Prism.Avalonia.Toolkit.csproj", "{95D737A5-18AA-4397-A8A3-B2E293ACE97E}"
+EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Source", "Source", "{70B9D432-E8BD-4EB9-9CA8-E63E2C804F3F}"
 	ProjectSection(SolutionItems) = preProject
 		README.md = README.md
@@ -118,6 +120,10 @@ Global
 		{97F8C6D3-308C-40BD-B181-5CDFCF101CBF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{97F8C6D3-308C-40BD-B181-5CDFCF101CBF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{97F8C6D3-308C-40BD-B181-5CDFCF101CBF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{95D737A5-18AA-4397-A8A3-B2E293ACE97E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{95D737A5-18AA-4397-A8A3-B2E293ACE97E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{95D737A5-18AA-4397-A8A3-B2E293ACE97E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{95D737A5-18AA-4397-A8A3-B2E293ACE97E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -140,6 +146,7 @@ Global
 		{340538B6-0781-470E-8BD1-7B7735FDC554} = {966DDDE9-D31A-404B-9F24-5378DE8A12C8}
 		{DAFC499B-E379-4453-BDCE-2EC30E1290C8} = {966DDDE9-D31A-404B-9F24-5378DE8A12C8}
 		{97F8C6D3-308C-40BD-B181-5CDFCF101CBF} = {5F8D330E-F611-4E17-8217-F2D6CDB8E1F7}
+		{95D737A5-18AA-4397-A8A3-B2E293ACE97E} = {9FF880E7-B56A-4E8E-9226-90A928FDE3AB}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {D260344D-8648-441D-B85B-9D280005F098}
